requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
barcode-generator==0.1rc15
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.4
cx_Oracle==8.3.0
distro==1.9.0
Django==5.0.2
django-auth-ldap==5.2.0
django-hCaptcha==0.2.0
django-ranged-response==0.2.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
fdb==2.0.2
firebirdsql==1.3.2
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.3
jiter==0.10.0
MarkupSafe==3.0.2
numpy==2.3.0
openai==1.88.0
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
oracledb==3.1.1
pandas==2.3.0
pdf2image==1.17.0
pdfkit==1.0.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-ldap @ file:///C:/Users/gustavo.nunes/Downloads/python_ldap-3.4.4-cp312-cp312-win_amd64.whl#sha256=6532815a3367b3ba3018e2f93c74916684c041fc6e9b441a1007c6e4e1034b48
pytz==2025.2
pyzbar==0.1.9
qrcode==7.4.2
reportlab==4.0.9
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
waitress==3.0.2
